pyproject.toml
[project]
name = "dlt-airflow-lineage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=2.10",
    "dlt[sql_database,filesystem,duckdb]>=1.6.0",
    "pymysql>=1.1.1",
]

[tool.uv]
dev-dependencies = [
    "setuptools>=75.8",
    "ruff>=0.9",
]
